make activity model dependant instead of object dependant
